feat: added bluez docker image

+++ scripts/install.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-BLEZ_VERSION="v0.2.0"
+BLEZ_VERSION="2022-12-05"
 DEFAULT_BLUEZ_VERSION="${BLUEZ_VERSION:-5.66}"
 
 PARENT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
@@ -32,9 +32,9 @@ function platform {
 function install {
     BLUEZ_VERSION="$1"
     BLUEZ_DIR="bluez-$BLUEZ_VERSION"
-    BLUEZ_ARTIFACT="$BLUEZ_DIR-$(platform).tar.gz"
     BLUEZ_HOME="/opt/$BLUEZ_DIR"
-    URL="https://github.com/charbonnierg/blez/releases/tag/$BLEZ_VERSION/bluez-$BLUEZ_VERSION-$(platform).tar.gz"
+    BLUEZ_ARTIFACT="$BLUEZ_DIR-$(platform).tar.gz"
+    URL="https://github.com/charbonnierg/blez/releases/tag/$BLEZ_VERSION/$BLUEZ_ARTIFACT"
     # Download build artifact from URL
     curl -O -q "$URL"
     # Uncompress artifact
@@ -53,8 +53,10 @@ function install {
     # Remove etc dir once files are used
     rm -rf "$BLUEZ_HOME/etc"
     # Add bluez directories to PATH through profile
-    cp $PARENT_DIR/profile.d/bluetooth.sh /etc/profile.d/bluetooth.sh
-    sed -i "s@__MARKER__@$BLUEZ_HOME@" /etc/profile.d/bluetooth.sh
+    if [ ! -f "/etc/profile.d/bluetooth.sh" ]; then
+        cp $PARENT_DIR/profile.d/bluetooth.sh /etc/profile.d/bluetooth.sh
+        sed -i "s@__MARKER__@$BLUEZ_HOME@" /etc/profile.d/bluetooth.sh
+    fi
 }
 
-install $1
+install "${1:-DEFAULT_BLUEZ_VERSION}"
